+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 - Added an option to disable speaking any text OCR -> eanbled, set to False
 - Add realm sound for emblems on the ground
 - Add wardrobe castle sound
-- Add blacksmith sound effect
+- Add blacksmith castle sound effect
+- Add enchanter castle sound effect
 - UI screen
   - anointment claim from false god